Replace VirtualDesktopAccessor with pyvda (#736)
* Update virtual desktop functions

* Update virtual desktop rules

* Add pyvda dependency

* Remove old dll

* Move rule to navigation

* add import exception checking
