+++ docs/LICENSE.md
@@ -178,4 +178,3 @@ Below is a list of all the non-original code used in this work, and the licenses
 - pywin32: Python Software Foundation License
 - Pillow: PIL Software License
 - TiRG by trg787: Public Domain License, see (github.com/synkarius/tirg-dll)
-- VirtualDesktopAccessor by Ciantic: MIT License
