+++ requirements-dev.txt
@@ -7,3 +7,4 @@ mock==3.0.5
 pylint==1.9.5
 appdirs==1.4.3
 scandir==1.10.0
+pyvda
