+++ requirements.txt
@@ -6,3 +6,4 @@ future==0.18.2
 mock==3.0.5
 appdirs==1.4.3
 scandir==1.10.0
+pyvda
